removed my-app directory scrolling done
