+++ src/frontend/json-server.json
@@ -1,0 +1,5 @@
+{
+    "port": 3333,
+    "delay": 1000,
+    "watch": "db.json"
+}
