+++ src/frontend/src/App.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import logo from './logo.svg';
 import './App.css';
 import Login from './components/Login';
 import Header from './components/Header';
