+++ src/frontend/src/sections/MainPageLeftSection.tsx
